pp2/TSIS/TSIS4/requirements.txt
pygame-ce>=2.4.0
psycopg2-binary>=2.9.0
